Rename mail_draft -> switchboard
