+++ switchboard/apps.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals
 from django.apps import AppConfig
 
 
-class MailerConfig(AppConfig):
-    name = 'mail_draft'
+class SwitchboardConfig(AppConfig):
+    name = 'switchboard'
